elimination of LogPage and Main Page comments
