+++ src/containers/MainPage/Header/styled.js
@@ -1,6 +1,5 @@
 import styled from 'styled-components';
 
-
 export const Header = styled.header`
 height: 50px;
 background: var(--Underpaper, #F6F6F6);
@@ -51,6 +50,7 @@ border: 1px solid var(--borders-elements, rgba(0, 0, 0, 0.20));
 display: inline-flex;
 margin-right: 30px;
 `
+
 export const SubTitle = styled.p`
 color: var(--Text-Black, #0D0D0D);
 font-family: SF Pro Text;
@@ -68,52 +68,4 @@ cursor: pointer;
 
 export const Colon = styled(SubTitle)`
 color: var(--borders-elements, rgba(0, 0, 0, 0.20));
-`
-
-export const HistoryTrack = styled.section`
-height: 50px;
-background: var(--Underpaper, #F6F6F6);
-display: flex;
-align-items: center;
-flex-direction: row;
-justify-content: space-between;
-border-top: 2px solid var(--borders-elements, rgba(0, 0, 0, 0.20));
-border-bottom: 2px solid var(--borders-elements, rgba(0, 0, 0, 0.20));
-`
-
-export const HistoryTrackLeftSide = styled(HistoryTrack)`
-justify-content: flex-start;
-border-top: 2px solid var(--borders-elements, rgba(0, 0, 0, 0.20));
-border-bottom: 2px solid var(--borders-elements, rgba(0, 0, 0, 0.20));
-`
-
-export const HistoryTrackRightSide = styled(HistoryTrackLeftSide)`
-border-left: 2px solid var(--borders-elements, rgba(0, 0, 0, 0.20));
-`
-
-export const ServerItem = styled.div`
-display: flex;
-padding: 5px 10px;
-align-items: center;
-gap: 5px;
-border-radius: 5px;
-box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.10);
-margin-left: 10px;
-`
-
-export const StatusImg = styled.img`
-margin: 0;
-`
-
-export const PointsItem = styled(StatusImg)`
-cursor: pointer;
-`
-
-export const StatusName = styled(SubTitle)`
-margin: 0 5px;
-`
-
-export const Cross = styled.img`
-margin: 0 14px;
-cursor: pointer;
 `
